api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
openai==1.42.0
faiss-cpu==1.8.0.post1
pypdf==4.3.1
python-dotenv==1.0.1
numpy==1.26.4
requests==2.32.3
